refactor: move `Database` imports to `src.dependencies.database`, clean up unused imports
