+++ src/models/guilds/guild.py
@@ -1,7 +1,7 @@
 from pydantic import Field
-from src.database import Database
+from src.dependencies.database import Database
 
-from src.utils.base import BaseModel, BaseList, optional_model
+from src.utils.base import BaseModel, optional_model
 from src.utils.errors import BadRequest400, NotFound404
