+++ src/models/guilds/leaderboards.py
@@ -5,7 +5,7 @@ from pydantic import BaseModel, Field
 
 from sanic_ext import openapi
 
-from src.database import Database
+from src.dependencies.database import Database
 
 
 @openapi.component()
