+++ src/models/projects/members.py
@@ -1,9 +1,4 @@
-from datetime import date, datetime
-
-from pydantic import Field
-from typing_extensions import Optional, Literal
-
-from src.database import Database
+from src.dependencies.database import Database
 from src.models.users import user
 from src.utils.base import BaseList
