+++ src/models/quests/objective.py
@@ -7,7 +7,7 @@ from src.models.quests.objective_customization.customization import Customizatio
 from src.models.quests.objective_targets.target import Targets
 from src.utils.base import BaseModel, BaseList, optional_model
 
-from src.database import Database
+from src.dependencies.database import Database
 
 from src.models.quests.reward import RewardCreateModel, RewardModel, RewardsListModel
