+++ src/models/server/items.py
@@ -3,7 +3,7 @@ from typing import Annotated, Optional
 from pydantic import Field, StringConstraints
 from sanic_ext import openapi
 
-from src.database import Database
+from src.dependencies.database import Database
 from src.utils.base import BaseModel, BaseList
 from src.utils.errors import BadRequest400, NotFound404
