+++ src/models/server/world.py
@@ -1,7 +1,7 @@
 from pydantic import Field
 from sanic_ext import openapi
 
-from src.database import Database
+from src.dependencies.database import Database
 from src.utils.base import BaseModel, optional_model
 from src.utils.errors import BadRequest400, NotFound404
