+++ src/models/users/interactions.py
@@ -3,7 +3,7 @@ import asyncio
 from pydantic import Field
 from fastapi import HTTPException
 
-from src.database import Database
+from src.dependencies.database import Database
 from src.utils.base import BaseModel, BaseList
