+++ src/models/users/user.py
@@ -4,9 +4,9 @@ from pydantic import Field
 from typing_extensions import Optional
 from fastapi import HTTPException
 
-from src.database import Database
+from src.dependencies.database import Database
 from src.models.users.profile import ProfileCreateModel, ProfileModel
-from src.utils.base import BaseModel, BaseList, optional_model
+from src.utils.base import BaseModel, optional_model
 
 
 class UserBaseModel(BaseModel):
