+++ src/routes/quest_progress.py
@@ -1,5 +1,3 @@
-from datetime import datetime
-
 from sanic import Blueprint, Request
 import sanic
 from sanic_ext import openapi, validate
@@ -7,7 +5,7 @@ from sanic_ext.extensions.openapi.definitions import Parameter, RequestBody, Res
 
 from src.utils.errors import BadRequest400, NotFound404
 
-from src.database import Database
+from src.dependencies.database import Database
 
 from src.models.quests import quest_progress, objective_progress
